fix: 64-bit integer to avoid overflow, collision checking + thread safety

+++ apple/RNCWebViewDecisionManager.h
@@ -5,16 +5,15 @@
 typedef void (^DecisionBlock)(BOOL);
 
 @interface RNCWebViewDecisionManager : NSObject {
-    int nextLockIdentifier;
+    NSInteger nextLockIdentifier;
     NSMutableDictionary *decisionHandlers;
 }
 
-@property (nonatomic) int nextLockIdentifier;
+@property (nonatomic) NSInteger nextLockIdentifier;
 @property (nonatomic, retain) NSMutableDictionary *decisionHandlers;
 
-+ (id)      getInstance;
++ (id)getInstance;
 
-- (int)setDecisionHandler:(DecisionBlock)handler;
-- (void)    setResult:(BOOL)shouldStart
-    forLockIdentifier:(int)lockIdentifier;
+- (NSInteger)setDecisionHandler:(DecisionBlock)handler;
+- (void)setResult:(BOOL)shouldStart forLockIdentifier:(NSInteger)lockIdentifier;
 @end
